echopage/requirements.txt
beautifulsoup4
requests
edge-tts
python-dotenv
ffmpeg-python
click
